create shared state and move stuff around
