+++ src/lib/components/objects/object.types.ts
@@ -1,6 +1,6 @@
 export type Component = {
   base: string;
-  shadow: string;
+  shadow?: string;
   light: string;
   name: string;
   rotation?: number;
