+++ src/lib/components/objects/square.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-    import { getLight, getShadow } from '../utils/mixers';
-    import { type Component } from './component.types';
+    import { getLight, getShadow } from '$lib/utils/mixers';
+    import { type Component } from './object.types';
 
     let { light, base, rotation = 0 }: Component = $props();
