+++ src/lib/store/config.types.ts
@@ -1,0 +1,12 @@
+export type Config = {
+  sunlight: string,
+  sphere: {
+    base: string,
+  },
+  square: {
+    base: string,
+    rotation: number,
+  },
+  skyBase: string,
+  groundBase: string,
+};
